@@ -1,8 +1,8 @@
 using GeoJSON.Net.CoordinateReferenceSystem;
 using GeoJSON.Net.Feature;
 using GeoJSON.Net.Geometry;
-using Newtonsoft.Json;
 using NUnit.Framework;
+using System.Text.Json;
 
 namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
 {
@@ -14,7 +14,7 @@ namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
         {
             var collection = new FeatureCollection();
 
-            var json = JsonConvert.SerializeObject(collection);
+            var json = JsonSerializer.Serialize(collection);
 
             Assert.IsTrue(!json.Contains("\"crs\""));
         }
@@ -23,8 +23,8 @@ namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
         public void Can_Deserialize_When_Json_Does_Not_Contain_Crs_Property()
         {
             var json = "{\"coordinates\":[90.65464646,53.2455662,200.4567],\"type\":\"Point\"}";
-
-            var point = JsonConvert.DeserializeObject<Point>(json);
+             
+            var point = JsonSerializer.Deserialize<Point>(json);
 
             Assert.IsNull(point.CRS);
         }
@@ -34,7 +34,7 @@ namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
         {
             var json = "{\"coordinates\": [ 90.65464646, 53.2455662, 200.4567 ], \"type\": \"Point\", \"crs\": { \"type\": \"name\", \"properties\": { \"name\": \"urn:ogc:def:crs:OGC:1.3:CRS84\" }}}";
 
-            var point = JsonConvert.DeserializeObject<Point>(json);
+            var point = JsonSerializer.Deserialize<Point>(json);
 
             Assert.IsNotNull(point.CRS);
             Assert.AreEqual(CRSType.Name, point.CRS.Type);
@@ -47,7 +47,7 @@ namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
                 "{\"type\":\"Point\",\"coordinates\":[34.56,12.34],\"crs\":{\"properties\":{\"name\":\"TEST NAME\"},\"type\":\"name\"}}";
             var point = new Point(new Position(12.34, 34.56)) { CRS = new NamedCRS("TEST NAME") };
 
-            var json = JsonConvert.SerializeObject(point);
+            var json = JsonSerializer.Serialize(point);
 
             Assert.IsNotNull(json);
             Assert.AreEqual(expected, json);
@@ -60,7 +60,7 @@ namespace GeoJSON.Net.Tests.CoordinateReferenceSystem
                 "{\"type\":\"Point\",\"coordinates\":[34.56,12.34],\"crs\":{\"properties\":{\"name\":\"urn:ogc:def:crs:OGC::CRS84\"},\"type\":\"name\"}}";
             var point = new Point(new Position(12.34, 34.56)) { CRS = new NamedCRS("urn:ogc:def:crs:OGC::CRS84") };
 
-            var json = JsonConvert.SerializeObject(point);
+            var json = JsonSerializer.Serialize(point);
 
             Assert.IsNotNull(json);
             Assert.AreEqual(expected, json);
