@@ -1,8 +1,9 @@
 using System;
 using System.Collections.Generic;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using GeoJSON.Net.Converters;
 using GeoJSON.Net.Geometry;
-using Newtonsoft.Json;
 using NUnit.Framework;
 
 namespace GeoJSON.Net.Tests.Geometry
@@ -116,9 +117,9 @@ namespace GeoJSON.Net.Tests.Geometry
         [TestCaseSource(typeof(GeometryTests), nameof(Geometries))]
         public void Can_Serialize_And_Deserialize_Geometry(IGeometryObject geometry)
         {
-            var json = JsonConvert.SerializeObject(geometry);
+            var json = JsonSerializer.Serialize(geometry);
 
-            var deserializedGeometry = JsonConvert.DeserializeObject<IGeometryObject>(json, new GeometryConverter());
+            var deserializedGeometry = JsonSerializer.Deserialize<IGeometryObject>(json);
 
             Assert.AreEqual(geometry, deserializedGeometry);
         }
@@ -127,7 +128,8 @@ namespace GeoJSON.Net.Tests.Geometry
         [TestCaseSource(typeof(GeometryTests), nameof(Geometries))]
         public void Serialization_Observes_Indenting_Setting_Of_Serializer(IGeometryObject geometry)
         {
-            var json = JsonConvert.SerializeObject(geometry, Formatting.Indented);
+            var options = new JsonSerializerOptions { WriteIndented = true };
+            var json = JsonSerializer.Serialize(geometry, options);
             Assert.IsTrue(json.Contains(Environment.NewLine));
         }
 
@@ -135,7 +137,7 @@ namespace GeoJSON.Net.Tests.Geometry
         [TestCaseSource(typeof(GeometryTests), nameof(Geometries))]
         public void Serialization_Observes_No_Indenting_Setting_Of_Serializer(IGeometryObject geometry)
         {
-            var json = JsonConvert.SerializeObject(geometry, Formatting.None);
+            var json = JsonSerializer.Serialize(geometry);
             Assert.IsFalse(json.Contains(Environment.NewLine));
             Assert.IsFalse(json.Contains(" "));
         }
@@ -146,9 +148,9 @@ namespace GeoJSON.Net.Tests.Geometry
         {
             var classWithGeometry = new ClassWithGeometryProperty(geometry);
 
-            var json = JsonConvert.SerializeObject(classWithGeometry);
+            var json = JsonSerializer.Serialize(classWithGeometry);
 
-            var deserializedClassWithGeometry = JsonConvert.DeserializeObject<ClassWithGeometryProperty>(json);
+            var deserializedClassWithGeometry = JsonSerializer.Deserialize<ClassWithGeometryProperty>(json);
 
             Assert.AreEqual(classWithGeometry, deserializedClassWithGeometry);
         }
@@ -159,13 +161,13 @@ namespace GeoJSON.Net.Tests.Geometry
         {
             var classWithGeometry = new ClassWithGeometryProperty(geometry);
 
-            var json = JsonConvert.SerializeObject(classWithGeometry);
+            var json = JsonSerializer.Serialize(classWithGeometry);
 
-            var deserializedClassWithGeometry = JsonConvert.DeserializeObject<ClassWithGeometryProperty>(json);
+            var deserializedClassWithGeometry = JsonSerializer.Deserialize<ClassWithGeometryProperty>(json);
 
             var actual = classWithGeometry;
             var expected = deserializedClassWithGeometry;
-            
+
             Assert.IsTrue(actual.Equals(expected));
             Assert.IsTrue(actual.Equals(actual));
 
