@@ -1,20 +1,22 @@
 ﻿// Copyright © Joerg Battermann 2014, Matt Hunt 2017
 
+using GeoJSON.Net.Geometry;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
-using GeoJSON.Net.Geometry;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
+using System.Text;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace GeoJSON.Net.Converters
 {
     /// <summary>
     /// Converter to read and write the <see cref="IEnumerable{MultiPolygon}" /> type.
     /// </summary>
-    public class PolygonEnumerableConverter : JsonConverter
+    public class PolygonEnumerableConverter : JsonConverter<IEnumerable<Polygon>>
     {
-        
+
         private static readonly LineStringEnumerableConverter PolygonConverter = new LineStringEnumerableConverter();
         /// <summary>
         ///     Determines whether this instance can convert the specified object type.
@@ -25,7 +27,7 @@ namespace GeoJSON.Net.Converters
         /// </returns>
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(IEnumerable<Polygon>);
+            return typeof(IEnumerable<Polygon>).IsAssignableFromType(objectType);
         }
 
         /// <summary>
@@ -38,19 +40,49 @@ namespace GeoJSON.Net.Converters
         /// <returns>
         ///     The object value.
         /// </returns>
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        public override IEnumerable<Polygon> Read(
+            ref Utf8JsonReader reader,
+            Type type,
+            JsonSerializerOptions options)
         {
-            var o = serializer.Deserialize<JArray>(reader);
-            var polygons =
-                o.Select(
-                    polygonObject => (IEnumerable<LineString>) PolygonConverter.ReadJson(
-                            polygonObject.CreateReader(),
-                            typeof(IEnumerable<LineString>),
-                            polygonObject, serializer))
-                    .Select(lines => new Polygon(lines))
-                    .ToList();
+            JsonElement polygons;
+            switch (reader.TokenType)
+            {
+                case JsonTokenType.Null:
+                    return null;
+                case JsonTokenType.StartArray:
+                    polygons = JsonDocument.ParseValue(ref reader).RootElement;
+                    break;
+                default:
+                    throw new InvalidOperationException("Incorrect json type");
+            }
 
-            return polygons;
+            var enumerator = polygons.EnumerateArray();
+            var lineStrings = new List<IEnumerable<LineString>>();
+            int i = 0;
+            while (enumerator.MoveNext())
+            {
+                var element = enumerator.Current;
+                byte[] bytes = Encoding.UTF8.GetBytes(element.GetRawText());
+                var stream = new MemoryStream(bytes);
+
+                var buffer = new byte[stream.Length];
+
+                // Fill the buffer.
+                // For this snippet, we're assuming the stream is open and has data.
+                // If it might be closed or empty, check if the return value is 0.
+                stream.Read(buffer, 0, (int)stream.Length);
+
+                var elementReader = new Utf8JsonReader(buffer, isFinalBlock: false, state: default);
+
+                lineStrings.Add(PolygonConverter.Read(
+                            ref elementReader,
+                            typeof(IEnumerable<LineString>),
+                            options));
+                i++;
+            }
+
+            return lineStrings.Select(lines => new Polygon(lines));
         }
 
         /// <summary>
@@ -59,21 +91,18 @@ namespace GeoJSON.Net.Converters
         /// <param name="writer">The <see cref="T:Newtonsoft.Json.JsonWriter" /> to write to.</param>
         /// <param name="value">The value.</param>
         /// <param name="serializer">The calling serializer.</param>
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+
+        public override void Write(
+            Utf8JsonWriter writer,
+            IEnumerable<Polygon> value,
+            JsonSerializerOptions options)
         {
-            if (value is IEnumerable<Polygon> polygons)
+            writer.WriteStartArray();
+            foreach (var polygon in value)
             {
-                writer.WriteStartArray();
-                foreach (var polygon in polygons)
-                {
-                    PolygonConverter.WriteJson(writer, polygon.Coordinates, serializer);
-                }
-                writer.WriteEndArray();
+                PolygonConverter.Write(writer, polygon.Coordinates, options);
             }
-            else
-            {
-                throw new ArgumentException($"{nameof(PointEnumerableConverter)}: unsupported value {value}");
-            }
+            writer.WriteEndArray();
         }
     }
 }
