@@ -1,17 +1,19 @@
 ﻿// Copyright © Joerg Battermann 2014, Matt Hunt 2017
 
 using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using GeoJSON.Net.CoordinateReferenceSystem;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
 namespace GeoJSON.Net.Converters
 {
     /// <summary>
     /// Converts <see cref="ICRSObject"/> types to and from JSON.
     /// </summary>
-    public class CrsConverter : JsonConverter
+    public class CrsConverter : JsonConverter<object>
     {
+        public override bool HandleNull => true;
+
         /// <summary>
         /// Determines whether this instance can convert the specified object type.
         /// </summary>
@@ -39,55 +41,69 @@ namespace GeoJSON.Net.Converters
         ///     or
         /// CRS must have a "type" property
         /// </exception>
-        public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+        public override object Read(
+            ref Utf8JsonReader reader,
+            Type type,
+            JsonSerializerOptions options)
         {
-            if (reader.TokenType == JsonToken.Null)
+            if (reader.TokenType == JsonTokenType.Null)
             {
                 return new UnspecifiedCRS();
             }
-            if (reader.TokenType != JsonToken.StartObject)
+            if (reader.TokenType != JsonTokenType.StartObject)
             {
-                throw new JsonReaderException("CRS must be null or a json object");
+                throw new JsonException("CRS must be null or a json object");
             }
 
-            var jObject = JObject.Load(reader);
-
-            JToken token;
-            if (!jObject.TryGetValue("type", StringComparison.OrdinalIgnoreCase, out token))
+            var jObject = JsonDocument.ParseValue(ref reader).RootElement;
+            JsonElement token;
+            if (!jObject.TryGetProperty("type", out token))
             {
-                throw new JsonReaderException("CRS must have a \"type\" property");
+                throw new JsonException("CRS must have a \"type\" property");
             }
 
-            var crsType = token.Value<string>();
+            var crsType = token.GetString();
 
             if (string.Equals("name", crsType, StringComparison.OrdinalIgnoreCase))
             {
-                JObject properties = null;
-                if (jObject.TryGetValue("properties", out token))
+                if (jObject.TryGetProperty("properties", out var properties))
                 {
-                    properties = token as JObject;
-                }
+                    var name = properties.GetProperty("name").GetString();
 
-                if (properties != null)
-                {
-                    var target = new NamedCRS(properties["name"].ToString());
-                    serializer.Populate(jObject.CreateReader(), target);
+                    var target = new NamedCRS(name);
+                    //var converted = JsonSerializer.Deserialize<NamedCRS>(properties);
+                    var converted = JsonSerializer.Deserialize<NamedCRS>(ref reader);
+
+                    if (converted.Properties != null)
+                    {
+                        foreach (var item in converted?.Properties)
+                        {
+                            target.Properties[item.Key] = item.Value;
+                        }
+                    }
+
                     return target;
                 }
             }
             else if (string.Equals("link", crsType, StringComparison.OrdinalIgnoreCase))
             {
-                JObject properties = null;
-                if (jObject.TryGetValue("properties", out token))
+                if (jObject.TryGetProperty("properties", out var properties))
                 {
-                    properties = token as JObject;
-                }
+                    var href = properties.GetProperty("href").GetString();
 
-                if (properties != null)
-                {
-                    var linked = new LinkedCRS(properties["href"].ToString());
-                    serializer.Populate(jObject.CreateReader(), linked);
-                    return linked;
+                    var target = new LinkedCRS(href);
+                    //var converted = JsonSerializer.Deserialize<LinkedCRS>(properties);
+                    var converted = JsonSerializer.Deserialize<LinkedCRS>(ref reader);
+
+                    if (converted.Properties != null)
+                    {
+                        foreach (var item in converted?.Properties)
+                        {
+                            target.Properties[item.Key] = item.Value;
+                        }
+                    }
+
+                    return target;
                 }
             }
 
@@ -101,20 +117,31 @@ namespace GeoJSON.Net.Converters
         /// <param name="value">The value.</param>
         /// <param name="serializer">The calling serializer.</param>
         /// <exception cref="System.ArgumentOutOfRangeException"></exception>
-        public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+        public override void Write(
+            Utf8JsonWriter writer,
+            object crsValue,
+            JsonSerializerOptions options)
         {
-            var crsObject = (ICRSObject)value;
+            var value = (ICRSObject)crsValue;
+            
+            if(value == null)
+                return;
 
-            switch (crsObject.Type)
+            switch (value.Type)
             {
                 case CRSType.Name:
-                    serializer.Serialize(writer, value);
+                    //var nameObject = (NamedCRS)value;
+                    //var serializedName = JsonSerializer.Serialize(nameObject, options);
+                    JsonSerializer.Serialize(writer, value, typeof(NamedCRS), options);
                     break;
                 case CRSType.Link:
-                    serializer.Serialize(writer, value);
+                    //var linkedObject = (LinkedCRS)value;
+                    //var serializedLink = JsonSerializer.Serialize(linkedObject, options);
+                    //writer.WriteRawValue(serializedLink);
+                    JsonSerializer.Serialize(writer, value, typeof(LinkedCRS), options);
                     break;
                 case CRSType.Unspecified:
-                    writer.WriteNull();
+                    writer.WriteNullValue();
                     break;
                 default:
                     throw new ArgumentOutOfRangeException();
