@@ -4,14 +4,14 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Text.Json.Serialization;
 using GeoJSON.Net.Converters;
-using Newtonsoft.Json;
 
 namespace GeoJSON.Net.Geometry
 {
     /// <summary>
     /// Defines the Polygon type.
-    /// Coordinates of a Polygon are a list of linear rings coordinate arrays. The first element in 
+    /// Coordinates of a Polygon are a list of linear rings coordinate arrays. The first element in
     /// the array represents the exterior ring. Any subsequent elements represent interior rings (or holes).
     /// </summary>
     /// <remarks>
@@ -23,7 +23,7 @@ namespace GeoJSON.Net.Geometry
         /// Initializes a new instance of the <see cref="Polygon" /> class.
         /// </summary>
         /// <param name="coordinates">
-        /// The linear rings with the first element in the array representing the exterior ring. 
+        /// The linear rings with the first element in the array representing the exterior ring.
         /// Any subsequent elements represent interior rings (or holes).
         /// </param>
         public Polygon(IEnumerable<LineString> coordinates)
@@ -36,7 +36,7 @@ namespace GeoJSON.Net.Geometry
                                             " (see GeoJSON spec at 'https://tools.ietf.org/html/rfc7946#section-3.1.6').", nameof(coordinates));
             }
 
-            
+
         }
 
         /// <summary>
@@ -50,12 +50,15 @@ namespace GeoJSON.Net.Geometry
         {
         }
 
+        [JsonPropertyName("type")]
+        //, Required = Required.Always, DefaultValueHandling = DefaultValueHandling.Include)]
+        [JsonConverter(typeof(JsonStringEnumConverter))]
         public override GeoJSONObjectType Type => GeoJSONObjectType.Polygon;
 
         /// <summary>
         /// Gets the list of linestrings defining this <see cref="Polygon" />.
         /// </summary>
-        [JsonProperty("coordinates", Required = Required.Always)]
+        [JsonPropertyName("coordinates")]
         [JsonConverter(typeof(LineStringEnumerableConverter))]
         public ReadOnlyCollection<LineString> Coordinates { get; }
 
